+++ src/stateManager.js
@@ -130,13 +130,13 @@ var StateManager = function(component, appCtx, initCtxObj) {
 		newState = newState || {};
 		newStateKeys = Object.keys(newState);
 
-		//Check to see if appState is a ready made state object. If so
+		//Check to see if appState is a ready made _state object. If so
 		//pass it straight to the stateChangeHandler. If a callback was passed in
 		//it would be assigned to newState
 		if(isControllerViewModel(newState)) {
 			willUndo = true;
 			nextState = extend(newState, staticState);
-			prevState = newState.previousState;
+			prevState = newState._previousState;
 			redoState = newAppState;
 			//Need to reset ViewModel with instance object so that setState is associated with
 			//the current ViewModel. This reason this ccurs is that when a ViewModel is created it
@@ -153,8 +153,8 @@ var StateManager = function(component, appCtx, initCtxObj) {
 				if(domain.hasOwnProperty(dataContext)){
 					for(linkedDataContext in links[dataContext]){
 						if(links[dataContext].hasOwnProperty(linkedDataContext)){
-							nextState[dataContext].state[links[dataContext][linkedDataContext]] =
-								(linkedDataContext in domain) ? extend(nextState[linkedDataContext].state) :
+							nextState[dataContext]._state[links[dataContext][linkedDataContext]] =
+								(linkedDataContext in domain) ? extend(nextState[linkedDataContext]._state) :
 									nextState[linkedDataContext];
 						}
 					}
@@ -256,15 +256,15 @@ var StateManager = function(component, appCtx, initCtxObj) {
 										if(subscriber in links){
 											for(dataContext in links[subscriber]){
 												if(links[subscriber].hasOwnProperty(dataContext)){
-													nextState[subscriber].state[links[subscriber][dataContext]] =
-														(dataContext in domain) ? extend(nextState[dataContext].state) :
+													nextState[subscriber]._state[links[subscriber][dataContext]] =
+														(dataContext in domain) ? extend(nextState[dataContext]._state) :
 															nextState[dataContext];
 												}
 												if(dataContext in links){
 													for(dataContext2 in links[dataContext]){
 														if(links[dataContext].hasOwnProperty(dataContext2)){
-															nextState[dataContext].state[links[dataContext][dataContext2]] =
-																(dataContext2 in domain) ? extend(nextState[dataContext2].state) :
+															nextState[dataContext]._state[links[dataContext][dataContext2]] =
+																(dataContext2 in domain) ? extend(nextState[dataContext2]._state) :
 																	nextState[dataContext2];
 														}
 													}
@@ -296,14 +296,14 @@ var StateManager = function(component, appCtx, initCtxObj) {
 					if(processedStateKeys[keyIdx] in links[namespace]){
 						for(dataContext in links[namespace][processedStateKeys[keyIdx]]){
 							if(links[namespace][processedStateKeys[keyIdx]].hasOwnProperty(dataContext)){
-								nextState[dataContext].state[links[namespace][processedStateKeys[keyIdx]][dataContext]] =
+								nextState[dataContext]._state[links[namespace][processedStateKeys[keyIdx]][dataContext]] =
 									nextState[processedStateKeys[keyIdx]];
 							}
 							if(dataContext in links){
 								for(dataContext2 in links[dataContext]){
 									if(links[dataContext].hasOwnProperty(dataContext2)){
-										nextState[dataContext].state[links[dataContext][dataContext2]] =
-											(dataContext2 in domain) ? extend(nextState[dataContext2].state) :
+										nextState[dataContext]._state[links[dataContext][dataContext2]] =
+											(dataContext2 in domain) ? extend(nextState[dataContext2]._state) :
 												nextState[dataContext2];
 									}
 								}
@@ -314,13 +314,13 @@ var StateManager = function(component, appCtx, initCtxObj) {
 					if(processedStateKeys[keyIdx] in links){
 						for(dataContext in links[processedStateKeys[keyIdx]]){
 							if(links[processedStateKeys[keyIdx]].hasOwnProperty(dataContext)){
-								nextState[processedStateKeys[keyIdx]].state[links[processedStateKeys[keyIdx]][dataContext]] =
+								nextState[processedStateKeys[keyIdx]]._state[links[processedStateKeys[keyIdx]][dataContext]] =
 									nextState[dataContext];
 							}
 							if(dataContext in links){
 								for(dataContext2 in links[dataContext]){
 									if(links[dataContext].hasOwnProperty(dataContext2)){
-										nextState[dataContext].state[links[dataContext][dataContext2]] =
+										nextState[dataContext]._state[links[dataContext][dataContext2]] =
 											nextState[dataContext2];
 									}
 								}
@@ -331,13 +331,13 @@ var StateManager = function(component, appCtx, initCtxObj) {
 	    }
 
 			if(stateMgr.appState.canRevert && calledBack){
-				prevState = stateMgr.appState.previousState;
+				prevState = stateMgr.appState._previousState;
 			} else if(hasStatic && staticState._staticUpdated && staticState._onlyStatic){
         if(stateMgr.appState.canRevert){
-        	prevState = stateMgr.appState.previousState;
+        	prevState = stateMgr.appState._previousState;
         }
         if(stateMgr.appState.canAdvance){
-        	redoState = stateMgr.appState.nextState;
+        	redoState = stateMgr.appState._nextState;
         }
       } else {
 				prevState = stateMgr.appState;
@@ -354,7 +354,7 @@ var StateManager = function(component, appCtx, initCtxObj) {
 		try {
 			//Add dataContextWillUpdate
       if(willUndo){
-        nextState = extend(nextState, {dataContextWillUpdate: newState.state.dataContextWillUpdate});
+        nextState = extend(nextState, {dataContextWillUpdate: newState._state.dataContextWillUpdate});
       } else {
         nextState = extend(nextState, {dataContextWillUpdate: processedState});
       }
@@ -364,7 +364,7 @@ var StateManager = function(component, appCtx, initCtxObj) {
 			!external || nextState.pageNotFound, remember);	
 
 			Object.freeze(stateMgr.appState);
-			Object.freeze(stateMgr.appState.state);
+			Object.freeze(stateMgr.appState._state);
 
 			if(typeof callback === 'function'){
 				calledBack = true;
@@ -453,7 +453,7 @@ var StateManager = function(component, appCtx, initCtxObj) {
   try {
   	this.ApplicationDataContext = controllerViewModel.call(this, appStateChangeHandler.bind(this, namespace));
   	stateMgr.appState = new this.ApplicationDataContext(void(0), void(0), void(0), enableUndo, routingEnabled);
-    stateMgr.appState.state = stateMgr.appState.state || {};
+    stateMgr.appState._state = stateMgr.appState._state || {};
   } catch (e) { 
   	if (e instanceof TypeError) {
     	throw new TypeError('Please assign a ControllerViewModel to the "controllerViewModel" prop in React.renderComponent');
@@ -469,7 +469,7 @@ var StateManager = function(component, appCtx, initCtxObj) {
 	for(viewModel in domain){
 		if(domain.hasOwnProperty(viewModel)){
 			dataContexts[viewModel] = domain[viewModel].call(this, appStateChangeHandler.bind(this, viewModel));
-			stateMgr.appState.state[viewModel] = new dataContexts[viewModel](stateMgr.appState.state[viewModel], true);
+			stateMgr.appState._state[viewModel] = new dataContexts[viewModel](stateMgr.appState._state[viewModel], true);
     }
   }
 
@@ -523,7 +523,7 @@ var StateManager = function(component, appCtx, initCtxObj) {
 		if(domain.hasOwnProperty(viewModel)){
 			for(link in links[viewModel]){
 			  if(links[viewModel].hasOwnProperty(link)){
-					stateMgr.appState[viewModel].state[links[viewModel][link]] = stateMgr.appState[link];
+					stateMgr.appState[viewModel]._state[links[viewModel][link]] = stateMgr.appState[link];
 			  }
 			}
 		}
@@ -532,8 +532,8 @@ var StateManager = function(component, appCtx, initCtxObj) {
 	//reinitialize with all data in place
 	for(viewModel in domain){
 		if(domain.hasOwnProperty(viewModel)){
-			stateMgr.appState.state[viewModel] =
-				new dataContexts[viewModel](stateMgr.appState.state[viewModel]);
+			stateMgr.appState._state[viewModel] =
+				new dataContexts[viewModel](stateMgr.appState._state[viewModel]);
 
 			if('getRoutes' in stateMgr.appState[viewModel].constructor.originalSpec){
 				routeHash = stateMgr.appState[viewModel].constructor.originalSpec.getRoutes();
@@ -602,10 +602,10 @@ var StateManager = function(component, appCtx, initCtxObj) {
 	// hasStatic = stateMgr.appState.constructor.originalSpec.__processedSpec__.hasStatic;
 	hasStatic = stateMgr.appState.constructor.getDescriptor().hasStatic;
 	if(hasStatic){
-		staticState = updateStatic(stateMgr.appState.constructor.getDescriptor().statics, stateMgr.appState.state);
+		staticState = updateStatic(stateMgr.appState.constructor.getDescriptor().statics, stateMgr.appState._state);
 	}
 
-  Object.freeze(stateMgr.appState.state);
+  Object.freeze(stateMgr.appState._state);
   Object.freeze(stateMgr.appState);
 
   if('dataContextWillInitialize' in stateMgr.appState.constructor.originalSpec){
@@ -671,13 +671,13 @@ StateManager.prototype.dispose = function(){
 	//reset dataContexts
 	for(viewModel in viewModels){
 		if(viewModels.hasOwnProperty(viewModel)){
-			resetState[viewModel] = this.appState[viewModel].dispose();
+			resetState[viewModel] = this.appState[viewModel].__dispose();
     }
   }
   resetState = extend(this.appState, emptyState, resetState);
 	try {
   	this.appState = new this.ApplicationDataContext(resetState, void(0), void(0), false, false);
-    this.appState.state = this.appState.state || {};
+    this.appState._state = this.appState._state || {};
   } catch (e) { 
   	if (e instanceof TypeError) {
     	throw new TypeError('Something went wrong');